testing data persistence w songs
there should be some prints that show when a song is added and when the .txt file is saved to the current directory. To test, open up the app, add a song or two, it will show in console the resulting dataframe after saving. Then, close out the app, reopen, and look in console and try to add a new song.